@@ -1,23 +1,19 @@
-apply plugin: 'com.android.library'
-
-configurations {
-    upload
+plugins {
+    id 'com.android.library'
+    id 'maven-publish'
 }
 
-group = groupId
-version = "${baseVersion}"
-
 android {
-    compileSdkVersion 29
+    compileSdkVersion 30
 
     defaultConfig {
         minSdkVersion 19
-        targetSdkVersion 29
+        targetSdkVersion 30
         versionCode 1
         versionName version
 
         // Use the test runner with JUnit4 support
-        testInstrumentationRunner "android.support.test.runner.AndroidJUnitRunner"
+        testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
     }
     buildTypes {
         debug {
