@@ -1,6 +1,6 @@
 package net.gini.android;
 
-import static android.support.test.InstrumentationRegistry.getTargetContext;
+import static androidx.test.core.app.ApplicationProvider.getApplicationContext;
 
 import static net.gini.android.helpers.TrustKitHelper.resetTrustKit;
 
@@ -15,10 +15,11 @@ import android.content.res.AssetManager;
 import android.graphics.Bitmap;
 import android.graphics.BitmapFactory;
 import android.os.Build;
-import android.support.test.filters.LargeTest;
-import android.support.test.filters.SdkSuppress;
-import android.support.test.runner.AndroidJUnit4;
+import androidx.test.filters.LargeTest;
+import androidx.test.filters.SdkSuppress;
+import androidx.test.ext.junit.runners.AndroidJUnit4;
 import android.util.Log;
+import android.util.Pair;
 
 import com.android.volley.toolbox.NoCache;
 
@@ -60,7 +61,7 @@ public class SdkIntegrationTest {
 
     @Before
     public void setUp() throws Exception {
-        final AssetManager assetManager = getTargetContext().getResources().getAssets();
+        final AssetManager assetManager = getApplicationContext().getResources().getAssets();
         final InputStream testPropertiesInput = assetManager.open("test.properties");
         assertNotNull("test.properties not found", testPropertiesInput);
         final Properties testProperties = new Properties();
@@ -77,7 +78,7 @@ public class SdkIntegrationTest {
 
         resetTrustKit();
 
-        gini = new SdkBuilder(getTargetContext(), clientId, clientSecret, "example.com").
+        gini = new SdkBuilder(getApplicationContext(), clientId, clientSecret, "example.com").
                 setApiBaseUrl(apiUri).
                 setUserCenterApiBaseUrl(userCenterUri).
                 setConnectionTimeoutInMs(60000).
@@ -92,7 +93,7 @@ public class SdkIntegrationTest {
 
     @Test
     public void deprecatedProcessDocumentBitmap() throws IOException, InterruptedException, JSONException {
-        final AssetManager assetManager = getTargetContext().getResources().getAssets();
+        final AssetManager assetManager = getApplicationContext().getResources().getAssets();
         final InputStream testDocumentAsStream = assetManager.open("test.jpg");
         assertNotNull("test image test.jpg could not be loaded", testDocumentAsStream);
 
@@ -103,7 +104,7 @@ public class SdkIntegrationTest {
 
     @Test
     public void processDocumentBitmap() throws IOException, InterruptedException, JSONException {
-        final AssetManager assetManager = getTargetContext().getResources().getAssets();
+        final AssetManager assetManager = getApplicationContext().getResources().getAssets();
         final InputStream testDocumentAsStream = assetManager.open("test.jpg");
         assertNotNull("test image test.jpg could not be loaded", testDocumentAsStream);
 
@@ -115,7 +116,7 @@ public class SdkIntegrationTest {
 
     @Test
     public void processDocumentByteArray() throws IOException, InterruptedException, JSONException {
-        final AssetManager assetManager = getTargetContext().getResources().getAssets();
+        final AssetManager assetManager = getApplicationContext().getResources().getAssets();
         final InputStream testDocumentAsStream = assetManager.open("test.jpg");
         assertNotNull("test image test.jpg could not be loaded", testDocumentAsStream);
 
@@ -127,14 +128,14 @@ public class SdkIntegrationTest {
 
     @Test
     public void processDocumentWithCustomCache() throws IOException, JSONException, InterruptedException {
-        gini = new SdkBuilder(getTargetContext(), clientId, clientSecret, "example.com").
+        gini = new SdkBuilder(getApplicationContext(), clientId, clientSecret, "example.com").
                 setApiBaseUrl(apiUri).
                 setUserCenterApiBaseUrl(userCenterUri).
                 setConnectionTimeoutInMs(60000).
                 setCache(new NoCache()).
                 build();
 
-        final AssetManager assetManager = getTargetContext().getResources().getAssets();
+        final AssetManager assetManager = getApplicationContext().getResources().getAssets();
         final InputStream testDocumentAsStream = assetManager.open("test.jpg");
         assertNotNull("test image test.jpg could not be loaded", testDocumentAsStream);
 
@@ -146,7 +147,7 @@ public class SdkIntegrationTest {
 
     @Test
     public void sendFeedback() throws Exception {
-        final AssetManager assetManager = getTargetContext().getResources().getAssets();
+        final AssetManager assetManager = getApplicationContext().getResources().getAssets();
         final InputStream testDocumentAsStream = assetManager.open("test.jpg");
         assertNotNull("test image test.jpg could not be loaded", testDocumentAsStream);
 
@@ -178,8 +179,8 @@ public class SdkIntegrationTest {
     @Test
     public void documentUploadWorksAfterNewUserWasCreatedIfUserWasInvalid() throws IOException, JSONException, InterruptedException {
         EncryptedCredentialsStore credentialsStore = new EncryptedCredentialsStore(
-                getTargetContext().getSharedPreferences("GiniTests", Context.MODE_PRIVATE), getTargetContext());
-        gini = new SdkBuilder(getTargetContext(), clientId, clientSecret, "example.com").
+                getApplicationContext().getSharedPreferences("GiniTests", Context.MODE_PRIVATE), getApplicationContext());
+        gini = new SdkBuilder(getApplicationContext(), clientId, clientSecret, "example.com").
                 setApiBaseUrl(apiUri).
                 setUserCenterApiBaseUrl(userCenterUri).
                 setConnectionTimeoutInMs(60000).
@@ -190,7 +191,7 @@ public class SdkIntegrationTest {
         UserCredentials invalidUserCredentials = new UserCredentials("invalid@example.com", "1234");
         credentialsStore.storeUserCredentials(invalidUserCredentials);
 
-        final AssetManager assetManager = getTargetContext().getResources().getAssets();
+        final AssetManager assetManager = getApplicationContext().getResources().getAssets();
         final InputStream testDocumentAsStream = assetManager.open("test.jpg");
         assertNotNull("test image test.jpg could not be loaded", testDocumentAsStream);
 
@@ -207,15 +208,15 @@ public class SdkIntegrationTest {
     public void emailDomainIsUpdatedForExistingUserIfEmailDomainWasChanged() throws IOException, JSONException, InterruptedException {
         // Upload a document to make sure we have a valid user
         EncryptedCredentialsStore credentialsStore = new EncryptedCredentialsStore(
-                getTargetContext().getSharedPreferences("GiniTests", Context.MODE_PRIVATE), getTargetContext());
-        gini = new SdkBuilder(getTargetContext(), clientId, clientSecret, "example.com").
+                getApplicationContext().getSharedPreferences("GiniTests", Context.MODE_PRIVATE), getApplicationContext());
+        gini = new SdkBuilder(getApplicationContext(), clientId, clientSecret, "example.com").
                 setApiBaseUrl(apiUri).
                 setUserCenterApiBaseUrl(userCenterUri).
                 setConnectionTimeoutInMs(60000).
                 setCredentialsStore(credentialsStore).
                 build();
 
-        final AssetManager assetManager = getTargetContext().getResources().getAssets();
+        final AssetManager assetManager = getApplicationContext().getResources().getAssets();
         final InputStream testDocumentAsStream = assetManager.open("test.jpg");
         assertNotNull("test image test.jpg could not be loaded", testDocumentAsStream);
 
@@ -227,7 +228,7 @@ public class SdkIntegrationTest {
         // Create another sdk instance with a new email domain (to simulate an app update)
         // and verify that the new email domain is used
         String newEmailDomain = "beispiel.com";
-        gini = new SdkBuilder(getTargetContext(), clientId, clientSecret, newEmailDomain).
+        gini = new SdkBuilder(getApplicationContext(), clientId, clientSecret, newEmailDomain).
                 setApiBaseUrl(apiUri).
                 setUserCenterApiBaseUrl(userCenterUri).
                 setConnectionTimeoutInMs(60000).
@@ -243,13 +244,13 @@ public class SdkIntegrationTest {
     @Test
     public void publicKeyPinningWithMatchingPublicKey() throws Exception {
         resetTrustKit();
-        gini = new SdkBuilder(getTargetContext(), clientId, clientSecret, "example.com").
+        gini = new SdkBuilder(getApplicationContext(), clientId, clientSecret, "example.com").
                 setNetworkSecurityConfigResId(net.gini.android.test.R.xml.network_security_config).
                 setApiBaseUrl(apiUri).
                 setUserCenterApiBaseUrl(userCenterUri).
                 setConnectionTimeoutInMs(60000).
                 build();
-        final AssetManager assetManager = getTargetContext().getResources().getAssets();
+        final AssetManager assetManager = getApplicationContext().getResources().getAssets();
         final InputStream testDocumentAsStream = assetManager.open("test.jpg");
         assertNotNull("test image test.jpg could not be loaded", testDocumentAsStream);
 
@@ -262,7 +263,7 @@ public class SdkIntegrationTest {
     @Test
     public void publicKeyPinningWithCustomCache() throws Exception {
         resetTrustKit();
-        gini = new SdkBuilder(getTargetContext(), clientId, clientSecret, "example.com").
+        gini = new SdkBuilder(getApplicationContext(), clientId, clientSecret, "example.com").
                 setNetworkSecurityConfigResId(net.gini.android.test.R.xml.network_security_config).
                 setApiBaseUrl(apiUri).
                 setUserCenterApiBaseUrl(userCenterUri).
@@ -270,7 +271,7 @@ public class SdkIntegrationTest {
                 setCache(new NoCache()).
                 build();
 
-        final AssetManager assetManager = getTargetContext().getResources().getAssets();
+        final AssetManager assetManager = getApplicationContext().getResources().getAssets();
         final InputStream testDocumentAsStream = assetManager.open("test.jpg");
         assertNotNull("test image test.jpg could not be loaded", testDocumentAsStream);
 
@@ -284,14 +285,14 @@ public class SdkIntegrationTest {
     @SdkSuppress(maxSdkVersion = Build.VERSION_CODES.JELLY_BEAN_MR1)
     public void publicKeyPinningWithWrongPublicKey() throws Exception {
         resetTrustKit();
-        gini = new SdkBuilder(getTargetContext(), clientId, clientSecret, "example.com").
+        gini = new SdkBuilder(getApplicationContext(), clientId, clientSecret, "example.com").
                 setNetworkSecurityConfigResId(net.gini.android.test.R.xml.wrong_network_security_config).
                 setApiBaseUrl(apiUri).
                 setUserCenterApiBaseUrl(userCenterUri).
                 setConnectionTimeoutInMs(60000).
                 build();
 
-        final AssetManager assetManager = getTargetContext().getResources().getAssets();
+        final AssetManager assetManager = getApplicationContext().getResources().getAssets();
         final InputStream testDocumentAsStream = assetManager.open("test.jpg");
         assertNotNull("test image test.jpg could not be loaded", testDocumentAsStream);
 
@@ -329,14 +330,14 @@ public class SdkIntegrationTest {
     @SdkSuppress(maxSdkVersion = Build.VERSION_CODES.JELLY_BEAN_MR1)
     public void publicKeyPinningWithMultiplePublicKeys() throws Exception {
         resetTrustKit();
-        gini = new SdkBuilder(getTargetContext(), clientId, clientSecret, "example.com").
+        gini = new SdkBuilder(getApplicationContext(), clientId, clientSecret, "example.com").
                 setNetworkSecurityConfigResId(net.gini.android.test.R.xml.multiple_keys_network_security_config).
                 setApiBaseUrl(apiUri).
                 setUserCenterApiBaseUrl(userCenterUri).
                 setConnectionTimeoutInMs(60000).
                 build();
 
-        final AssetManager assetManager = getTargetContext().getResources().getAssets();
+        final AssetManager assetManager = getApplicationContext().getResources().getAssets();
         final InputStream testDocumentAsStream = assetManager.open("test.jpg");
         assertNotNull("test image test.jpg could not be loaded", testDocumentAsStream);
 
@@ -348,7 +349,7 @@ public class SdkIntegrationTest {
 
     @Test
     public void createPartialDocument() throws Exception {
-        final AssetManager assetManager = getTargetContext().getResources().getAssets();
+        final AssetManager assetManager = getApplicationContext().getResources().getAssets();
         final InputStream testDocumentAsStream = assetManager.open("multi-page-p1.png");
         assertNotNull("test image multi-page-p1.png could not be loaded", testDocumentAsStream);
 
@@ -364,7 +365,7 @@ public class SdkIntegrationTest {
 
     @Test
     public void deletePartialDocumentWithoutParents() throws Exception {
-        final AssetManager assetManager = getTargetContext().getResources().getAssets();
+        final AssetManager assetManager = getApplicationContext().getResources().getAssets();
         final InputStream testDocumentAsStream = assetManager.open("multi-page-p1.png");
         assertNotNull("test image multi-page-p1.png could not be loaded", testDocumentAsStream);
 
@@ -385,7 +386,7 @@ public class SdkIntegrationTest {
 
     @Test
     public void deletePartialDocumentWithParents() throws Exception {
-        final AssetManager assetManager = getTargetContext().getResources().getAssets();
+        final AssetManager assetManager = getApplicationContext().getResources().getAssets();
         final InputStream page1Stream = assetManager.open("multi-page-p1.png");
         assertNotNull("test image multi-page-p1.png could not be loaded", page1Stream);
 
@@ -416,7 +417,7 @@ public class SdkIntegrationTest {
 
     @Test
     public void deletePartialDocumentFailsWhenNotDeletingParents() throws Exception {
-        final AssetManager assetManager = getTargetContext().getResources().getAssets();
+        final AssetManager assetManager = getApplicationContext().getResources().getAssets();
         final InputStream page1Stream = assetManager.open("multi-page-p1.png");
         assertNotNull("test image multi-page-p1.png could not be loaded", page1Stream);
 
@@ -447,7 +448,7 @@ public class SdkIntegrationTest {
 
     @Test
     public void processCompositeDocument() throws Exception {
-        final AssetManager assetManager = getTargetContext().getResources().getAssets();
+        final AssetManager assetManager = getApplicationContext().getResources().getAssets();
         final InputStream page1Stream = assetManager.open("multi-page-p1.png");
         assertNotNull("test image multi-page-p1.png could not be loaded", page1Stream);
         final InputStream page2Stream = assetManager.open("multi-page-p2.png");
@@ -541,7 +542,7 @@ public class SdkIntegrationTest {
 
     @Test
     public void testDeleteCompositeDocument() throws Exception {
-        final AssetManager assetManager = getTargetContext().getResources().getAssets();
+        final AssetManager assetManager = getApplicationContext().getResources().getAssets();
         final InputStream page1Stream = assetManager.open("multi-page-p1.png");
         assertNotNull("test image multi-page-p1.png could not be loaded", page1Stream);
 
@@ -579,40 +580,43 @@ public class SdkIntegrationTest {
     }
 
     private Map<Document, Map<String, SpecificExtraction>> processDocument(DocumentUploadBuilder uploadBuilder)
-            throws InterruptedException, JSONException {
-        final DocumentTaskManager documentTaskManager = gini.getDocumentTaskManager();
+            throws InterruptedException {
+        final Pair<Document, Map<String, SpecificExtraction>> retrieveExtractions =
+                analyzeDocument(uploadBuilder, 10, 0);
 
-        final Task<Document> upload = uploadBuilder.upload(documentTaskManager);
-        final Task<Document> processDocument = upload.onSuccessTask(new Continuation<Document, Task<Document>>() {
-            @Override
-            public Task<Document> then(Task<Document> task) throws Exception {
-                Document document = task.getResult();
-                return documentTaskManager.pollDocument(document);
-            }
-        });
-
-        final Task<Map<String, SpecificExtraction>> retrieveExtractions = processDocument.onSuccessTask(
-                new Continuation<Document, Task<Map<String, SpecificExtraction>>>() {
-                    @Override
-                    public Task<Map<String, SpecificExtraction>> then(Task<Document> task) throws Exception {
-                        return documentTaskManager.getExtractions(task.getResult());
-                    }
-                });
-
-        retrieveExtractions.waitForCompletion();
-        if (retrieveExtractions.isFaulted()) {
-            Log.e("TEST", Log.getStackTraceString(retrieveExtractions.getError()));
-        }
-
-        assertFalse("extractions should have succeeded", retrieveExtractions.isFaulted());
-
-        final Map<String, SpecificExtraction> extractions = retrieveExtractions.getResult();
+        final Map<String, SpecificExtraction> extractions = retrieveExtractions.second;
 
         assertEquals("IBAN should be found", "DE78370501980020008850", extractions.get("iban").getValue());
         assertEquals("Amount to pay should be found", "1.00:EUR", extractions.get("amountToPay").getValue());
         assertEquals("BIC should be found", "COLSDE33", extractions.get("bic").getValue());
         assertEquals("Payee should be found", "Uno Flüchtlingshilfe", extractions.get("senderName").getValue());
 
-        return Collections.singletonMap(upload.getResult(), extractions);
+        return Collections.singletonMap(retrieveExtractions.first, extractions);
+    }
+
+    private Pair<Document, Map<String, SpecificExtraction>> analyzeDocument(DocumentUploadBuilder uploadBuilder, int maxRetries, int retries)
+            throws InterruptedException {
+        final DocumentTaskManager documentTaskManager = gini.getDocumentTaskManager();
+
+        final Task<Document> upload = uploadBuilder.upload(documentTaskManager);
+        final Task<Document> processDocument = upload.onSuccessTask(task -> {
+            Document document = task.getResult();
+            return documentTaskManager.pollDocument(document);
+        });
+
+        final Task<Map<String, SpecificExtraction>> retrieveExtractions = processDocument
+                .onSuccessTask(task -> documentTaskManager.getExtractions(task.getResult()));
+
+        retrieveExtractions.waitForCompletion();
+
+        if (retrieveExtractions.isFaulted() && retries < maxRetries) {
+            Log.e("TEST", Log.getStackTraceString(retrieveExtractions.getError()));
+            Thread.sleep(1000);
+            return analyzeDocument(uploadBuilder, maxRetries, retries + 1);
+        }
+
+        assertFalse("extractions should have succeeded", retrieveExtractions.isFaulted());
+
+        return new Pair<>(upload.getResult(), retrieveExtractions.getResult());
     }
 }
