@@ -1,6 +1,6 @@
 package net.gini.android;
 
-import static android.support.test.InstrumentationRegistry.getTargetContext;
+import static androidx.test.core.app.ApplicationProvider.getApplicationContext;
 
 import static net.gini.android.helpers.TrustKitHelper.resetTrustKit;
 
@@ -15,10 +15,11 @@ import android.content.res.AssetManager;
 import android.graphics.Bitmap;
 import android.graphics.BitmapFactory;
 import android.os.Build;
-import android.support.test.filters.LargeTest;
-import android.support.test.filters.SdkSuppress;
-import android.support.test.runner.AndroidJUnit4;
+import androidx.test.filters.LargeTest;
+import androidx.test.filters.SdkSuppress;
+import androidx.test.ext.junit.runners.AndroidJUnit4;
 import android.util.Log;
+import android.util.Pair;
 
 import com.android.volley.toolbox.NoCache;
 
@@ -57,7 +58,7 @@ public class SdkIntegrationTestAccounting {
 
     @Before
     public void setUp() throws Exception {
-        final AssetManager assetManager = getTargetContext().getResources().getAssets();
+        final AssetManager assetManager = getApplicationContext().getResources().getAssets();
         final InputStream testPropertiesInput = assetManager.open("test.properties");
         assertNotNull("test.properties not found", testPropertiesInput);
         final Properties testProperties = new Properties();
@@ -77,7 +78,7 @@ public class SdkIntegrationTestAccounting {
 
         resetTrustKit();
 
-        gini = new SdkBuilder(getTargetContext(), clientId, clientSecret, "example.com").
+        gini = new SdkBuilder(getApplicationContext(), clientId, clientSecret, "example.com").
                 setApiBaseUrl(apiUriAccounting).
                 setGiniApiType(GiniApiType.ACCOUNTING).
                 setUserCenterApiBaseUrl(userCenterUri).
@@ -116,7 +117,7 @@ public class SdkIntegrationTestAccounting {
 
     @Test
     public void processDocumentWithCustomCache() throws IOException, JSONException, InterruptedException {
-        gini = new SdkBuilder(getTargetContext(), clientId, clientSecret, "example.com").
+        gini = new SdkBuilder(getApplicationContext(), clientId, clientSecret, "example.com").
                 setApiBaseUrl(apiUriAccounting).
                 setGiniApiType(GiniApiType.ACCOUNTING).
                 setUserCenterApiBaseUrl(userCenterUri).
@@ -158,8 +159,8 @@ public class SdkIntegrationTestAccounting {
     @Test
     public void documentUploadWorksAfterNewUserWasCreatedIfUserWasInvalid() throws IOException, JSONException, InterruptedException {
         EncryptedCredentialsStore credentialsStore = new EncryptedCredentialsStore(
-                getTargetContext().getSharedPreferences("GiniTests", Context.MODE_PRIVATE), getTargetContext());
-        gini = new SdkBuilder(getTargetContext(), clientId, clientSecret, "example.com").
+                getApplicationContext().getSharedPreferences("GiniTests", Context.MODE_PRIVATE), getApplicationContext());
+        gini = new SdkBuilder(getApplicationContext(), clientId, clientSecret, "example.com").
                 setApiBaseUrl(apiUriAccounting).
                 setGiniApiType(GiniApiType.ACCOUNTING).
                 setUserCenterApiBaseUrl(userCenterUri).
@@ -184,8 +185,8 @@ public class SdkIntegrationTestAccounting {
     public void emailDomainIsUpdatedForExistingUserIfEmailDomainWasChanged() throws IOException, JSONException, InterruptedException {
         // Upload a document to make sure we have a valid user
         EncryptedCredentialsStore credentialsStore = new EncryptedCredentialsStore(
-                getTargetContext().getSharedPreferences("GiniTests", Context.MODE_PRIVATE), getTargetContext());
-        gini = new SdkBuilder(getTargetContext(), clientId, clientSecret, "example.com").
+                getApplicationContext().getSharedPreferences("GiniTests", Context.MODE_PRIVATE), getApplicationContext());
+        gini = new SdkBuilder(getApplicationContext(), clientId, clientSecret, "example.com").
                 setApiBaseUrl(apiUriAccounting).
                 setGiniApiType(GiniApiType.ACCOUNTING).
                 setUserCenterApiBaseUrl(userCenterUri).
@@ -201,7 +202,7 @@ public class SdkIntegrationTestAccounting {
         // Create another sdk instance with a new email domain (to simulate an app update)
         // and verify that the new email domain is used
         String newEmailDomain = "beispiel.com";
-        gini = new SdkBuilder(getTargetContext(), clientId, clientSecret, newEmailDomain).
+        gini = new SdkBuilder(getApplicationContext(), clientId, clientSecret, newEmailDomain).
                 setApiBaseUrl(apiUriAccounting).
                 setGiniApiType(GiniApiType.ACCOUNTING).
                 setUserCenterApiBaseUrl(userCenterUri).
@@ -217,7 +218,7 @@ public class SdkIntegrationTestAccounting {
 
     @Test
     public void publicKeyPinningWithMatchingPublicKey() throws Exception {
-        gini = new SdkBuilder(getTargetContext(), clientId, clientSecret, "example.com").
+        gini = new SdkBuilder(getApplicationContext(), clientId, clientSecret, "example.com").
                 setNetworkSecurityConfigResId(net.gini.android.test.R.xml.network_security_config).
                 setApiBaseUrl(apiUriAccounting).
                 setGiniApiType(GiniApiType.ACCOUNTING).
@@ -233,7 +234,7 @@ public class SdkIntegrationTestAccounting {
 
     @Test
     public void publicKeyPinningWithCustomCache() throws Exception {
-        gini = new SdkBuilder(getTargetContext(), clientId, clientSecret, "example.com").
+        gini = new SdkBuilder(getApplicationContext(), clientId, clientSecret, "example.com").
                 setNetworkSecurityConfigResId(net.gini.android.test.R.xml.network_security_config).
                 setApiBaseUrl(apiUriAccounting).
                 setGiniApiType(GiniApiType.ACCOUNTING).
@@ -251,7 +252,7 @@ public class SdkIntegrationTestAccounting {
     @Test
     @SdkSuppress(maxSdkVersion = Build.VERSION_CODES.JELLY_BEAN_MR1)
     public void publicKeyPinningWithWrongPublicKey() throws Exception {
-        gini = new SdkBuilder(getTargetContext(), clientId, clientSecret, "example.com").
+        gini = new SdkBuilder(getApplicationContext(), clientId, clientSecret, "example.com").
                 setNetworkSecurityConfigResId(net.gini.android.test.R.xml.wrong_network_security_config).
                 setApiBaseUrl(apiUriAccounting).
                 setGiniApiType(GiniApiType.ACCOUNTING).
@@ -292,7 +293,7 @@ public class SdkIntegrationTestAccounting {
     @Test
     @SdkSuppress(maxSdkVersion = Build.VERSION_CODES.JELLY_BEAN_MR1)
     public void publicKeyPinningWithMultiplePublicKeys() throws Exception {
-        gini = new SdkBuilder(getTargetContext(), clientId, clientSecret, "example.com").
+        gini = new SdkBuilder(getApplicationContext(), clientId, clientSecret, "example.com").
                 setNetworkSecurityConfigResId(net.gini.android.test.R.xml.multiple_keys_network_security_config).
                 setApiBaseUrl(apiUriAccounting).
                 setGiniApiType(GiniApiType.ACCOUNTING).
@@ -315,38 +316,41 @@ public class SdkIntegrationTestAccounting {
     }
 
     private Map<Document, Map<String, SpecificExtraction>> analyzeDocumentAndAssertExtractions(DocumentUploadBuilder uploadBuilder)
-            throws InterruptedException, JSONException {
-        final DocumentTaskManager documentTaskManager = gini.getDocumentTaskManager();
+            throws InterruptedException {
+        final Pair<Document, Map<String, SpecificExtraction>> retrieveExtractions =
+                analyzeDocument(uploadBuilder, 10, 0);
 
-        final Task<Document> upload = uploadBuilder.upload(documentTaskManager);
-        final Task<Document> processDocument = upload.onSuccessTask(new Continuation<Document, Task<Document>>() {
-            @Override
-            public Task<Document> then(Task<Document> task) throws Exception {
-                Document document = task.getResult();
-                return documentTaskManager.pollDocument(document);
-            }
-        });
-
-        final Task<Map<String, SpecificExtraction>> retrieveExtractions = processDocument.onSuccessTask(
-                new Continuation<Document, Task<Map<String, SpecificExtraction>>>() {
-                    @Override
-                    public Task<Map<String, SpecificExtraction>> then(Task<Document> task) throws Exception {
-                        return documentTaskManager.getExtractions(task.getResult());
-                    }
-                });
-
-        retrieveExtractions.waitForCompletion();
-        if (retrieveExtractions.isFaulted()) {
-            Log.e("TEST", Log.getStackTraceString(retrieveExtractions.getError()));
-        }
-
-        assertFalse("extractions should have succeeded", retrieveExtractions.isFaulted());
-
-        final Map<String, SpecificExtraction> extractions = retrieveExtractions.getResult();
+        final Map<String, SpecificExtraction> extractions = retrieveExtractions.second;
 
         assertEquals("Amount to pay should be found", "1.00:EUR", extractions.get("amountToPay").getValue());
         assertEquals("Payee should be found", "Uno Flüchtlingshilfe", extractions.get("senderName").getValue());
 
-        return Collections.singletonMap(upload.getResult(), extractions);
+        return Collections.singletonMap(retrieveExtractions.first, extractions);
+    }
+
+    private Pair<Document, Map<String, SpecificExtraction>> analyzeDocument(DocumentUploadBuilder uploadBuilder, int maxRetries, int retries)
+            throws InterruptedException {
+        final DocumentTaskManager documentTaskManager = gini.getDocumentTaskManager();
+
+        final Task<Document> upload = uploadBuilder.upload(documentTaskManager);
+        final Task<Document> processDocument = upload.onSuccessTask(task -> {
+            Document document = task.getResult();
+            return documentTaskManager.pollDocument(document);
+        });
+
+        final Task<Map<String, SpecificExtraction>> retrieveExtractions = processDocument
+                .onSuccessTask(task -> documentTaskManager.getExtractions(task.getResult()));
+
+        retrieveExtractions.waitForCompletion();
+
+        if (retrieveExtractions.isFaulted() && retries < maxRetries) {
+            Log.e("TEST", Log.getStackTraceString(retrieveExtractions.getError()));
+            Thread.sleep(1000);
+            return analyzeDocument(uploadBuilder, maxRetries, retries + 1);
+        }
+
+        assertFalse("extractions should have succeeded", retrieveExtractions.isFaulted());
+
+        return new Pair<>(upload.getResult(), retrieveExtractions.getResult());
     }
 }
