@@ -2,7 +2,7 @@ package net.gini.android.authorization.requests;
 
 import static org.junit.Assert.assertEquals;
 
-import android.support.test.runner.AndroidJUnit4;
+import androidx.test.ext.junit.runners.AndroidJUnit4;
 
 import com.android.volley.AuthFailureError;
 import com.android.volley.DefaultRetryPolicy;
