@@ -1,8 +1,13 @@
 package net.gini.android.authorization;
 
+import static com.android.volley.Request.Method.GET;
+import static com.android.volley.Request.Method.POST;
+import static com.android.volley.Request.Method.PUT;
+
 import android.net.Uri;
 
 import com.android.volley.RequestQueue;
+import com.android.volley.VolleyError;
 import com.android.volley.toolbox.JsonObjectRequest;
 
 import net.gini.android.GiniApiType;
@@ -15,14 +20,13 @@ import net.gini.android.requests.RetryPolicyFactory;
 import org.json.JSONException;
 import org.json.JSONObject;
 
+import java.io.UnsupportedEncodingException;
+import java.net.URLEncoder;
+import java.nio.charset.StandardCharsets;
 import java.util.HashMap;
 
 import bolts.Continuation;
 import bolts.Task;
-
-import static com.android.volley.Request.Method.GET;
-import static com.android.volley.Request.Method.POST;
-import static com.android.volley.Request.Method.PUT;
 
 
 /**
@@ -146,10 +150,15 @@ public class UserCenterAPICommunicator {
     // Visible for testing
     Task<JSONObject> getGiniApiSessionTokenInfo(final Session giniApiSession) {
         final RequestTaskCompletionSource<JSONObject> completionSource = RequestTaskCompletionSource.newCompletionSource();
-        final String url = mBaseUrl + "oauth/check_token?token=" + giniApiSession.getAccessToken();
-        final JsonObjectRequest request =
-                new JsonObjectRequest(GET, url, null, completionSource, completionSource);
-        mRequestQueue.add(request);
+        try {
+            final String url = mBaseUrl + "oauth/check_token?token=" +
+                    URLEncoder.encode(giniApiSession.getAccessToken(), "UTF-8");
+            final JsonObjectRequest request =
+                    new JsonObjectRequest(GET, url, null, completionSource, completionSource);
+            mRequestQueue.add(request);
+        } catch (UnsupportedEncodingException e) {
+            completionSource.onErrorResponse(new VolleyError(e));
+        }
         return completionSource.getTask();
     }
 
